requirements.txt
# ── Core scientific stack ───────────────────────────────────────────────────
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0
scikit-image>=0.21.0

# ── Computer Vision ─────────────────────────────────────────────────────────
opencv-python>=4.8.0
Pillow>=10.0.0

# ── Deep learning / CLIP ────────────────────────────────────────────────────
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0          # HuggingFace — hosts Fashion-CLIP / openai-clip
open-clip-torch>=2.23.0       # OpenCLIP: drop-in for Fashion-CLIP

# ── Data handling ────────────────────────────────────────────────────────────
pandas>=2.0.0
tqdm>=4.66.0
h5py>=3.9.0                   # efficient HDF5 feature cache on disk
joblib>=1.3.0                 # model serialisation (SVM) and scaler persistence

# ── Configuration & logging ──────────────────────────────────────────────────
pyyaml>=6.0
loguru>=0.7.0

# ── Visualisation (optional, used in notebooks) ──────────────────────────────
matplotlib>=3.7.0
seaborn>=0.12.0

# ── Testing ──────────────────────────────────────────────────────────────────
pytest>=7.4.0
pytest-cov>=4.1.0             # coverage reporting
